lexer: refactor the  regex pattern
So, I guess I over-engineer this...
Anyway, the idea is that each patter is defined individually as a constexpr
variable, we then put the patterns into an array, in the order that they
need to be evaluated.
Then another function, turns that array of patterns into an array of chars
that contains all the patterns and the separators.

The last piece is there are a few helper functions to get the position of a
pattern in the array of patterns. This is used in the if chain that find which
patterns was matched against.
Hopefully, this will make it easy to switch the order of the patterns without
having to change too much code... Hopefully...