Split the regex matcher into two matchers
CTRE, takes the regex as a template parameter.
The approach of having a single matcher was almost at
the limit for the fold expressions that clang processes.

The split puts, the identifier and the constant regexes
into a new matcher. This is probably the one that will grow
in the future. As it will need to handle other contant types.