Rework parts of the lexer to use named captures
The identifiers and constant lexer uses names captures instead
of using the positional ids.
This makes it easier to use subgroups in the regex.

Also, stop using the while match and use only the intended group.
This became an issue after changing the delimiter for the end of
a constant.