WIP
Attempt to change the type checker to change the AST for
compound assignment.

The idea was to implement the following transformation
origin:
int *f;
*f += 42;

transformation:
int *tmp_var = &(*f);
*tmp_var = *tmp_var + 42;

This way, the left expression is only evaluated once.
When assigning it to tmp_var;

However, this is not working.
I decided to change it to a compoung_statement.
Which breaks places where it's expecting an expression.

On second though, I think I need to implement the comma operator
and then turn this into a comma operation
something like this.
tmp_var = &(*f), *tmp_var = *tmp_var + 42;